% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnw2RmdPandoc.R
\name{Rnw2RmdPandoc}
\alias{Rnw2RmdPandoc}
\title{Translation from '.Rnw' to '.Rmd' using pandoc and custom
filters}
\usage{
Rnw2RmdPandoc(input, output = NULL, ..., options = NULL)
}
\arguments{
\item{input}{character(1) path to an \code{.Rnw} (Sweave) file.}

\item{output}{character(1) path to the output file}

\item{...}{additional arguments passed to \link{pandoc_convert}.}

\item{options}{character vector of options (in the style of
\link{system}, with key/value pairs as separate elements of the
character vector) passed to \link{pandoc_convert}. \code{Rnw2RmdPandoc()}
adds filters to translate LaTeX commands introduced by
BiocStyle.}
}
\description{
\code{Rnw2RmdPandoc()} uses pandoc and custom filters to
convert from LaTeX (Sweave) to markdown.
}
\details{
\code{Rnw2RmdPandoc()} requires that \code{pandoc} is
installed. Conversion is far from perfect; the more complicated
the LaTeX, the less perfect the conversion. Current major
limitations include:

\itemize{
\item The generated Rmd file does not include front-matter 'yaml',
e.g., title, author, and output format.
\item Cross-references are not handled.
}

Closely examine the generated markdown file for
\verb{\{=latex\}}; this indicates LaTeX code that has not be translated.
}
\examples{
rnw <- system.file(package = "Rnw2Rmd", "example", "Rnw2Rmd-test.Rnw")
cat(readLines(rnw), sep = "\n")

rmd <- Rnw2RmdPandoc(rnw, tempfile(fileext = ".Rmd"), verbose = TRUE)
cat(readLines(rmd), sep = "\n")

}
