% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnwToRmd.R
\name{RnwToRmd}
\alias{RnwToRmd}
\title{Knitr to R markdown converter}
\usage{
RnwToRmd(input)
}
\arguments{
\item{input}{name of file to be converted}
}
\value{
None
}
\description{
RnwToRmd converts a .Rnw file to .Rmd
}
\details{
The conversion involves three steps:
- insert the knitr R code into LaTex "verbatim"environments, so that the Rnw file is now pure LaTex code
- use pandoc to convert the intermediate LaTeX file to a markdown file
- convert the knitr code chunks to R markdown
}
\examples{
\dontrun{
RnwToRmd("test.Rnw")
}

}
\author{
Patrick Henaff, \email{pa.henaff@gmail.com}
}
\keyword{knitr,}
\keyword{rmarkdown}
